move sample tests over
